refactor: un mismo card para los pedidos
