--- src/features/seller/components/AssignOrdersPanel.tsx
@@ -1,5 +1,0 @@
-import RejectedOrdersPanel from './RejectedOrdersPanel';
-
-export default function AssignOrdersPanel(props: { embedded?: boolean }) {
-  return <RejectedOrdersPanel embedded={props.embedded} />;
-}
