+++ src/features/seller/components/CardOrder.tsx
@@ -1,0 +1,107 @@
+import type { Order } from '../types';
+import { formatCurrency } from '../utils/currency';
+import { formatDate } from '../utils/formatDate';
+import { StatusPill } from './StatusPill';
+
+interface Props {
+  order: Order;
+  onViewDetails: (order: Order) => void;
+  onPrimaryAction: (order: Order) => void;
+  primaryActionLabel: string;
+  isPrimaryActionLoading: boolean;
+  isPrimaryActionSuccess?: boolean;
+  successLabel?: string;
+  primaryActionLoadingLabel?: string;
+}
+
+export const CardOrder = ({
+  order,
+  onViewDetails,
+  onPrimaryAction,
+  primaryActionLabel,
+  isPrimaryActionLoading,
+  isPrimaryActionSuccess = false,
+  successLabel = 'Pedido marcado como listo',
+  primaryActionLoadingLabel = 'Procesando…',
+}: Props) => {
+  return (
+    <article className="overflow-hidden rounded-3xl border border-(--theme-border) bg-linear-to-br from-(--theme-card-bg) to-(--theme-secondary-bg)/40 shadow-sm duration-200 hover:shadow-xl">
+      <div className="p-5">
+        <div className="flex items-start justify-between gap-4">
+          <div className="min-w-0">
+            <h3 className="mt-1 text-xl font-900 tracking-tight text-(--theme-text)" style={{ fontFamily: 'Outfit, sans-serif' }}>
+              #{order.orderId.slice(-10).toUpperCase()}
+            </h3>
+            <p className="mt-2 text-sm font-700 text-(--theme-text) opacity-80">
+              {order.buyerName ?? 'Comprador desconocido'}
+            </p>
+          </div>
+
+          <div className="text-right">
+            <p className="text-[10px] uppercase tracking-[0.2em] text-(--theme-text) opacity-40">
+              Total
+            </p>
+            <p className="font-900 text-2xl tracking-tight text-primary">
+              {formatCurrency(order.total)}
+            </p>
+          </div>
+        </div>
+
+        <div className="mt-4 flex flex-wrap items-center gap-2">
+          <StatusPill status={order.status} />
+        </div>
+
+        <div className="mt-4 grid gap-3 sm:grid-cols-2">
+          <div className="rounded-2xl border border-(--theme-border) bg-(--theme-secondary-bg)/60 px-4 py-3">
+            <p className="text-[11px] font-800 uppercase tracking-[0.22em] text-(--theme-text) opacity-45">
+              Ubicación
+            </p>
+            <p className="mt-1 text-sm font-700 text-(--theme-text)">
+              {order.locationLabel ?? 'No registrada'}
+            </p>
+            <p className="mt-1 text-xs text-(--theme-text) opacity-55">
+              {order.locationType ?? 'Tipo no registrado'}
+            </p>
+          </div>
+
+          <div className="rounded-2xl border border-(--theme-border) bg-(--theme-secondary-bg)/60 px-4 py-3">
+            <p className="text-[11px] font-800 uppercase tracking-[0.22em] text-(--theme-text) opacity-45">
+              Fecha
+            </p>
+            <p className="mt-1 text-sm font-700 text-(--theme-text)">
+              {formatDate(order.createdAt)}
+            </p>
+            <p className="mt-1 text-xs text-(--theme-text) opacity-55">
+              Actualizado {formatDate(order.updatedAt)}
+            </p>
+          </div>
+        </div>
+
+        <div className="mt-5 flex flex-col gap-3 sm:flex-row sm:justify-end">
+          <button
+            type="button"
+            onClick={() => onViewDetails(order)}
+            className="inline-flex items-center justify-center rounded-full border border-(--theme-border) px-5 py-2.5 text-sm font-700 text-(--theme-text) transition hover:border-primary hover:bg-(--theme-secondary-bg) hover:text-primary"
+          >
+            Ver detalles
+          </button>
+
+          <button
+            type="button"
+            onClick={() => onPrimaryAction(order)}
+            disabled={isPrimaryActionLoading}
+            className="inline-flex items-center justify-center cursor-pointer rounded-full bg-primary px-5 py-2.5 text-sm font-800 text-white transition hover:opacity-90 disabled:opacity-50"
+          >
+            {isPrimaryActionLoading ? primaryActionLoadingLabel : primaryActionLabel}
+          </button>
+        </div>
+
+        {isPrimaryActionSuccess && (
+          <p className="mt-4 text-sm font-700 text-primary">
+            {successLabel}
+          </p>
+        )}
+      </div>
+    </article>
+  );
+};
