--- src/pages/seller/assign.astro
@@ -1,8 +1,0 @@
----
-import SellerLayout from '../../layouts/SellerLayout.astro';
-import AssignOrdersPanel from '../../features/seller/components/AssignOrdersPanel';
----
-
-<SellerLayout title="Asignar mensajeros">
-  <AssignOrdersPanel client:load embedded />
-</SellerLayout>
