--- src/pages/seller/orders.astro
@@ -1,8 +1,0 @@
----
-import SellerOrdersPanel from '../../features/seller/components/SellerOrderPanel';
-import SellerLayout from '../../layouts/SellerLayout.astro';
----
-
-<SellerLayout title="Pedidos empaquetados">
-  <SellerOrdersPanel client:load embedded />
-</SellerLayout>
